services/consumer/requirements.txt
kafka-python==2.0.2
flake8==3.9.0
python-dotenv==0.19.0
minio